Module siro::vdom

Representation of DOM nodes.

Structs

Map
A virtual node created by map.

Enums

AttributeValue
The value of DOM attributes.

Traits

Attributes
A collection of DOM attributes.

AttributesRenderer
A rendering context restricted to DOM attributes.

Event
An abstraction of DOM events.

EventDecoder
Decoder of DOM events.

Nodes
A collection of virtual DOM nodes.

NodesRenderer
The context for rendering a Nodes.

Functions

attribute
Create an Attr that specifies an arbitrary attribute value, like domNode.setAttribute(name, value).

class
Create an Attr that specify a CSS class name.

event
Create an Attr that registers an event.

inner_html
Create an Attr that specifies the inner HTML content of the element.

iter
Create a Nodes from an iterator.

property
Create an Attr that specifies an arbitrary property value, like domNode.name = value.

style
Create an Attr that specify an inline style.

Type Definitions

CowStr
Clone-on-write string.