Trait siro::vdom::AttributesRenderer

pub trait AttributesRenderer {
    type Msg: 'static;
    type Ok;
    type Error;
    fn attribute(
        &mut self, 
        name: CowStr, 
        value: AttributeValue
    ) -> Result<(), Self::Error>;

    fn property<T>(&mut self, name: CowStr, value: T) -> Result<(), Self::Error>
    where
        T: Serialize;

    fn event<D>(
        &mut self, 
        event_type: &'static str, 
        decoder: D
    ) -> Result<(), Self::Error>
    where
        D: EventDecoder<Msg = Self::Msg> + 'static;

    fn class(&mut self, class_name: CowStr) -> Result<(), Self::Error>;

    fn style(&mut self, name: CowStr, value: CowStr) -> Result<(), Self::Error>;

    fn inner_html(&mut self, inner_html: CowStr) -> Result<(), Self::Error>;

    fn end(self) -> Result<Self::Ok, Self::Error>;
}

A rendering context restricted to DOM attributes.

Associated Types

type Msg: 'static

type Ok

type Error

Loading content...
Required methods

fn attribute(
&mut self,
name: CowStr,
value: AttributeValue
) -> Result<(), Self::Error>

Add an attribute to this element, corresponding to domNode.setAttribute(name, value).

fn property<T>(&mut self, name: CowStr, value: T) -> Result<(), Self::Error> where
T: Serialize,

Add a property to this element, corresponding to domNode.name = value.

Currently, the property values are modeled by serde::Serialize and converted to JavaScript objects at rendering time.

fn event<D>(
&mut self,
event_type: &'static str,
decoder: D
) -> Result<(), Self::Error> where
D: EventDecoder<Msg = Self::Msg> + 'static,

Register an event callback to this element.

fn class(&mut self, class_name: CowStr) -> Result<(), Self::Error>

Add a class to this element.

fn style(&mut self, name: CowStr, value: CowStr) -> Result<(), Self::Error>

Apply an inline style to this element.

fn inner_html(&mut self, inner_html: CowStr) -> Result<(), Self::Error>

Set the content of inner HTML.

When this method is called, the additions of child nodes by append_child should be ignored.

fn end(self) -> Result<Self::Ok, Self::Error>

Complete the rendering of this element.

Loading content...